Media client testing
Former-commit-id: ea4bbf35624b68bb9f9631e65b128da89466e21a [formerly 92e3bbcec49e2922453344549595bd15d7faa331]
Former-commit-id: 23c7cfa6ed44d5d0b38f6dbf0c39eaf2a8d627e9

+++ api/domain/options.go
@@ -73,9 +73,9 @@ type (
 		MediaCompression     int        `json:"media_compression" binding:"required"`
 		MediaConvertWebP     bool       `json:"media_convert_webp"`
 		MediaServeWebP       bool       `json:"media_serve_webp"`
-		MediaUploadMaxSize   int        `json:"media_upload_max_size" binding:"numeric"`
-		MediaUploadMaxWidth  int        `json:"media_upload_max_width" binding:"numeric"`
-		MediaUploadMaxHeight int        `json:"media_upload_max_height" binding:"numeric"`
+		MediaUploadMaxSize   int64      `json:"media_upload_max_size" binding:"numeric"`
+		MediaUploadMaxWidth  int64      `json:"media_upload_max_width" binding:"numeric"`
+		MediaUploadMaxHeight int64      `json:"media_upload_max_height" binding:"numeric"`
 		MediaOrganiseDate    bool       `json:"media_organise_year_month"`
 		MediaSizes           MediaSizes `json:"media_images_sizes"`
 		// Server Cache
